ClickRemovalBase in lib-builtin-effects
