+++ src/effects/ClickRemoval.h
@@ -6,74 +6,18 @@
 
   Craig DeForest
 
-  (Structure largely stolen from NoiseRemoval.h by Dominic Mazzoni)
-
-  This file is intended to become part of Audacity.  You may modify and/or
-  distribute it under the same terms as Audacity itself.
-
 **********************************************************************/
 
 #ifndef __AUDACITY_EFFECT_CLICK_REMOVAL__
 #define __AUDACITY_EFFECT_CLICK_REMOVAL__
 
-#include "StatefulEffect.h"
+#include "ClickRemovalBase.h"
 #include "StatefulEffectUIServices.h"
-#include "ShuttleAutomation.h"
 #include <wx/weakref.h>
 
 class wxSlider;
 class wxTextCtrl;
-class Envelope;
 class ShuttleGui;
-class WaveChannel;
-
-class ClickRemovalBase : public StatefulEffect
-{
-public:
-   static inline ClickRemovalBase *
-   FetchParameters(ClickRemovalBase &e, EffectSettings &) { return &e; }
-   static const ComponentInterfaceSymbol Symbol;
-
-   ClickRemovalBase();
-   virtual ~ClickRemovalBase();
-
-   // ComponentInterface implementation
-
-   ComponentInterfaceSymbol GetSymbol() const override;
-   TranslatableString GetDescription() const override;
-   ManualPageID ManualPage() const override;
-
-   // EffectDefinitionInterface implementation
-
-   EffectType GetType() const override;
-
-   // Effect implementation
-
-   bool CheckWhetherSkipEffect(const EffectSettings &settings) const override;
-   bool Process(EffectInstance &instance, EffectSettings &settings) override;
-
-private:
-   bool ProcessOne(int count, WaveChannel &track,
-      sampleCount start, sampleCount len);
-
-   bool RemoveClicks(size_t len, float *buffer);
-
-protected:
-   Envelope *mEnvelope;
-
-   bool mbDidSomething; // This effect usually does nothing on real-world data.
-   size_t windowSize;
-   int mThresholdLevel;
-   int mClickWidth;
-   int sep;
-
-   const EffectParameterMethods& Parameters() const override;
-
-static constexpr EffectParameter Threshold{ &ClickRemovalBase::mThresholdLevel,
-   L"Threshold",  200,     0,       900,     1  };
-static constexpr EffectParameter Width{ &ClickRemovalBase::mClickWidth,
-   L"Width",      20,      0,       40,      1  };
-};
 
 class EffectClickRemoval final :
     public ClickRemovalBase,
